+++ packages/ui/src/dataviews/resource-data-view/state/use-selection.ts
@@ -1,0 +1,10 @@
+import { useState } from 'react';
+
+export function useSelection() {
+	const [selection, setSelection] = useState<string[]>([]);
+
+	return {
+		selection,
+		handleSelectionChange: setSelection,
+	};
+}